[RSDK-10413] create a conan.lock file (#28)
This was generated by running `conan lock create .`. I don't like that it doesn't have a newline at the end of the file, but I also don't like manually changing machine-generated files. :shrug: 

The ticket was about investigating whether this is a good idea, but I'm generally in favor of pinning specific versions of dependencies, so I just went ahead and tried it out.

Everything still builds on Linux, though I haven't tried out the Windows or Mac builds. but I've got another ticket to set up a CI, and when that happens, it'll be obvious if there are issues.


* create conan.lock

* make module.tar.gz should work, and the conan build artifacts should not be tracked