bugfix: get meta.json in the spot where we upload the module (#33)
Sigh. The two lines I couldn't test (invoking the Viam CLI) are failing because they can't find meta.json, because the code was built on a different runner than we're using to upload to the registry. Like #30, I don't have a way to try this out without merging, so still can't tell if it's right. but hopefully it'll work this time...

* bugfix: get meta.json in the spot where we upload the module

* bugfix from #32: fix the caching key