Don't modify meta.json in the Windows build (#35)
Per Friday's discussion, Windows is smart enough to figure out that `build-conan/build/RelWithDebInfo/tflite_cpu` really means `build-conan/build/RelWithDebInfo/tflite_cpu.exe`, with a `.exe` on the end. So, let's get rid of all the effort we put into having different `meta.json`s on different platforms. This also prevents us running into bugs like the one Abe and I found Friday, when there is a mismatch between the global meta.json and the local one. (Maybe we'll still need to fix such bugs in the future, but at least it won't be for this module.)


* don't modify meta.json on Windows: that OS can append '.exe' to the end of the entrypoint automatically

* no need to install jq any more