Pin the Go implementation for Debian Bullseye (#38)
I gave up on https://github.com/viam-modules/mlmodel-tflite/pull/37, so this is a bandaid: every time we build a new version of the repo, copy the old Go code and reupload that as the "new" version for Debian Bullseye only.

FWIW, the .tar.gz file with the old Go code does have a giant warning (both in the top-level directory and the directory that contains the entrypoint) explaining that this is not the C++ implementation.

This PR also fixes the checks to allow successful PRs to be merged again. Thanks to James O for help with that!


* REVERTME: try it out without merging

* re-upload the old version for Bullseye

* tweak comment

* Revert "REVERTME: try it out without merging"

This reverts commit 342664216cced15aa5012007b91faed7a68c2cbc.

* enable the bullseye upload, too

* trigger the github actions again

* Revert "trigger the github actions again"

This reverts commit dcaafac59c066c43fa1f6fb32398d197fb1e1d7e.

* Attempt to fix the CI checks: have a single action we depend on

* Adapting check from RDK's test.yml

* debugging

* comment tweak